Add Part 2 of Day01
Added a convert_to_digit function that converts the written numbers
into digits. This preserves the first and last letters of each word,
to avoid issues where the text would overlap, and so converting one
removes an earlier or later digit.

This is likely not always safe, and a sufficiently evil input may be
able to break it, but it works here.