Implement Day17 Part 1
Implemented day 17 part 1 using a graph representation. To handle the
restrictions on how many times it can travel in each direction before needing
to turn, graph embeds direction of travel and number of steps in that direction
in the nodes of the graph, and so has 12 layers of the basic map, 3 layers for
each direction of travel. Edge connections are then used to enforce the
movement restrictions.